Mudança para double - ponto flutuante com maior precisão

+++ include/config.hpp
@@ -12,12 +12,12 @@ class Config {
   size_t  count;
   float   damping;
 
-  int   length_1, length_2;
-  float mass_1, mass_2;
-  float initial_theta_1, initial_theta_2;
-  float final_theta_1, final_theta_2;
-  Color initial_color_1, initial_color_2;
-  Color final_color_1, final_color_2;
+  int    length_1, length_2;
+  float  mass_1, mass_2;
+  double initial_theta_1, initial_theta_2;
+  double final_theta_1, final_theta_2;
+  Color  initial_color_1, initial_color_2;
+  Color  final_color_1, final_color_2;
 
   bool        show_fps, debug_mode, paused, show_timer;
   std::string font_path;
